added lotto changes and api changes for lotto and company

+++ API/Models/Tbllotteryno.cs
@@ -19,7 +19,7 @@ public partial class Tbllotteryno
 
     public decimal? Paid { get; set; }
 
-    public ulong? LotteryStatus { get; set; }
+    public int? LotteryStatus { get; set; }
 
     public string? LotteryReferenceId { get; set; }
